Introduce Embeddable Commenting
This change introduces an embed endpoint, which is fetched and included
in the blog by
https://github.com/thoughtbot/robots.thoughtbot.com/blob/cp-include-carnival/source/javascripts/comments.js

This embed endpoint returns JavaScript (implemented in CoffeeScript) and
CSS (implemented in Julius) for the commenting functionality.

Currently the url is hardcoded for development, I've added issue #11 for
this.

We're not currently including jQuery, this script currently does not
work in IE (a very small fraction of our readers) and is more verbose
than it otherwise would be. I wanted to do this in order to not
introduce another dependency that will slow loading of the blog, and
also to see whether it's worth it.

There are currently no tests for the CoffeeScript. I wanted to progress
without the necessary infrastructure for testing in place, but I'd like
to write test. In the spirit of Carnival, I'm toying with the idea of
re-writing the CoffeeScript in TypeScript or something else as well, but
that would be something for the future...