Add user_name and (unimplemented) gravatar_url
- Introduce a UserComment type with the appropriate JSON instance
- Lookup the user for each comment and build a UserComment
- Return the list of UserComment

Notes:

The type is called UserComment (because that reads better) but the
arguments are reversed (Entity Comment, User) because that was more
convenient for partial application. If the user cannot be found for a
Comment, it is omitted from the results. The query is currently N+1;
joins are cumbersome (but possible) in Yesod, so I propose we wait until
this becomes an issue before rewriting.