Add custom Session module
Read the session key from an environment variable if present. This
should get sessions working across dynos since both dynos will be using
the same key and the data itself is stored in the browser cookie.

I did some local testing and it seems to work as intended. Only when I
set a value in the environment variable do my sessions survive a process
restart.

There are some general details in the module doc, but to use it on
Heroku, I'm going to do:

```
% heroku config:set SESSION_SECRET \
  "$(hexdump -v -n "48" -e '1/1 "%02x"' /dev/urandom)"
```