Manually allow base version to vary slightly in frozen cabal.config
Specifying exact versions for library packages which can be installed via cabal
is useful since it ensures a valid set of dependencies, that all users will be
using the same exact versions, and any users without those versions will have
them automatically installed on-demand.

Specifying an exact version for base requires all developers be on the exact
same version of GHC. Since only the installed version of base can be used,
having a mismatch will cause any installation attempts to fail. Allowing the
patch level to vary should minimize the occurrence of this without any
negative effects.

There are two caveats:

- At this time, all developers must still coordinate minor version upgrades
- The change must be done manually anytime the cabal.config file is regenerated
  by `cabal freeze`

This is a conservative first attempt. In the future, we might also allow the
minor version to vary.