Install executable-only packages outside of the sandbox
The dependencies of our executable-only packages and our application seem to
conflict. We're unable to determine the root cause, but installing the
executable-only packages outside of the sandbox solves the issue.

This also means these packages can be reused across projects. If we assume there
aren't as strict version requirements, this should be a Good Thing.