Use Halcyon for building, even locally
Halcyon[1] is a tool for building and deploying Haskell projects. It
handles compiling GHC, Cabal, and a project sandbox. It uses S3 to share
build artifacts across builds provided the platform is the same. It
encourages using hard version constraints to ensure consistent,
repeatable builds. A fully cached build can complete in seconds.

Halcyon is used by the Haskell on Heroku[2] buildpack which we've been
using for some time. This change makes it so that Halcyon is used for
local development and CI.

The change involved the following:

- Add a .halcyon directory for various build-related settings
- Add a halcyon-build script which wraps the call to halcyon, provides
  some default options, and ensures S3-related env vars are set
- Simplifications to bin/setup
- Simplifications to travis.yml
- Changes to README.md

This has been tested to work on Travis (Ubuntu) and Arch Linux.

1: https://halcyon.sh/
2: https://haskellonheroku.com/