Add sandbox-pre-build-hook to fix shebang
The coffee-script package on Ubuntu ships a coffee executable with a
hard-coded shebang of /usr/bin/nodejs. Since Halcyon installs nodejs in
an alternate directory and relies on a modified $PATH to find it, the
shebang needs to use /usr/bin/env. This pre-build hook uses sed to fix
Ubuntu's packaging error.