Add commentSite attribute to Comment
This is phase 1 of 2. Here, all existing comments are given an incorrect
SiteId of 1. This is OK because we aren't using it for anything yet.
After deployment, I will manually update all existing comments to have
the correct SiteId. I can then follow up with Phase 2: actually rely on
commentSite.