Restructure test modules
- HLint fixes
- Add test dependencies to constraints
- Add hspec-expectations-lifted

This removes the need for Helpers.Assertions

- Put generic runDB/authenticateAs helpers into TestImport directly

This (the rename and function move) gets us closer to the default
scaffold.

- Add Factories module

This replaces Helpers.DB. All functions live in DB for composability and
functions used by only one Example were moved to be directly with it.

- Rename Helpers.Request to Y.Test.Extension

This will soon become a PR to upstream

- Move Helpers.Response directly into CommentsSpec

It is only used there.