Move window.opener.postMessage to default-layout.coffee
Now that visitors may be browsing Carnival itself, the "Ultimate
Destination" feature could kick in, meaning that users aren't
necessarily brought to the homepage after logging in.

When embedded, if a login doesn't bring the user to the homepage, the
child window opened for login won't be closed. Moving the
window.opener.postMessage call to occur on every page ensures that if
there is a window.opener, it will always receive the message.