Integration testing with Test.Hspec.WebDriver
Tests logging in as a known user and leaving a comment on the homepage.

Spec.hs uses CPP to determine if it should run the Integration spec or
use auto-discovery to run the normal suite -- this reduces the impact on
the cabal file while still allowing a runtime choice of which suite to
run.