Call withEnv in TestImport.withApp
In e7cc415f, loadEnv was moved out of makeFoundation and placed earlier
in the application boot process. This unintentionally meant it was no
longer invoked during tests as part of withApp. This was OK on CircleCI
where the environment variables are explicitly set, but breaks tests
locally where they are set in .env.