Delete demo comments after one week
Comments on the demo site are for the purpose of trying out the site. We
expect a lot of "test and "hello" type of comments, and the page will
look strange with hundreds of old test comments on it.

This commit adds new carnival-tasks executable to the distribution for
executing one-off tasks as subcommands. Subcommands are implemented as
Handlers.

Currently, the only subcommand is "delete-stale-comments," which deletes
stale comments from the demo site.