Add stack support
https://github.com/commercialhaskell/stack

This allows the following basic workflow:

- stack setup
- stack build
- stack test
- stack exec -- ghci foo.hs
- stack exec -- yesod devel

This addition does not conflict with any other development or deployment
approaches. In the future, we could move to docker-based Heroku deployments too,
which would cut the last tie to Halcyon.

Note: most of the extra-deps are because we have slightly older upper-bounds on
a few dependencies and so we need to pull those version in from Hackage.
Updating those upper bounds to allow the snapshot versions should be safe in
most --if not all-- cases, but would require testing to ensure there's no
breakage.